[TrimmableTypeMap][Tests] Split test helpers into separate files
- ScannerRunner.cs: TypeMapEntry/MethodEntry/TypeMethodGroup records + scanner logic
- TypeDataBuilder.cs: TypeComparisonData record + Cecil/peer type data builders
- MarshalMethodDiffHelper.cs: marshal method comparison logic
- ComparisonDiffHelper.cs: per-field type comparison helpers
- ScannerComparisonTests.Helpers.cs: partial class with assertion helpers + assembly path properties
- ScannerComparisonTests.cs: now contains only [Fact] methods
