+++ tests/Microsoft.Android.Sdk.TrimmableTypeMap.IntegrationTests/ScannerRunner.cs
@@ -1,0 +1,190 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Java.Interop.Tools.Cecil;
+using Microsoft.Build.Utilities;
+using Mono.Cecil;
+using Xamarin.Android.Tasks;
+
+namespace Microsoft.Android.Sdk.TrimmableTypeMap.IntegrationTests;
+
+record TypeMapEntry (string JavaName, string ManagedName, bool SkipInJavaToManaged);
+
+record MethodEntry (string JniName, string JniSignature, string? Connector);
+
+record TypeMethodGroup (string ManagedName, List<MethodEntry> Methods);
+
+static class ScannerRunner
+{
+	public static (List<TypeMapEntry> entries, Dictionary<string, List<TypeMethodGroup>> methodsByJavaName) RunLegacy (string assemblyPath)
+	{
+		var cache = new TypeDefinitionCache ();
+		var resolver = new DefaultAssemblyResolver ();
+		resolver.AddSearchDirectory (Path.GetDirectoryName (assemblyPath)!);
+
+		var runtimeDir = Path.GetDirectoryName (typeof (object).Assembly.Location);
+		if (runtimeDir != null) {
+			resolver.AddSearchDirectory (runtimeDir);
+		}
+
+		var readerParams = new ReaderParameters { AssemblyResolver = resolver };
+		using var assembly = AssemblyDefinition.ReadAssembly (assemblyPath, readerParams);
+
+		var scanner = new XAJavaTypeScanner (
+			Xamarin.Android.Tools.AndroidTargetArch.Arm64,
+			new TaskLoggingHelper (new MockBuildEngine (), "test"),
+			cache
+		);
+
+		var javaTypes = scanner.GetJavaTypes (assembly);
+		var (dataSets, _) = TypeMapCecilAdapter.GetDebugNativeEntries (
+			javaTypes, cache, needUniqueAssemblies: false
+		);
+
+		var entries = dataSets.JavaToManaged
+			.Select (e => new TypeMapEntry (e.JavaName, e.ManagedName, e.SkipInJavaToManaged))
+			.OrderBy (e => e.JavaName, StringComparer.Ordinal)
+			.ThenBy (e => e.ManagedName, StringComparer.Ordinal)
+			.ToList ();
+
+		var methodsByJavaName = new Dictionary<string, List<TypeMethodGroup>> ();
+		foreach (var typeDef in javaTypes) {
+			var javaName = GetCecilJavaName (typeDef);
+			if (javaName == null) {
+				continue;
+			}
+
+			var managedName = GetManagedName (typeDef);
+			var methods = ExtractMethodRegistrations (typeDef);
+
+			if (!methodsByJavaName.TryGetValue (javaName, out var groups)) {
+				groups = new List<TypeMethodGroup> ();
+				methodsByJavaName [javaName] = groups;
+			}
+
+			groups.Add (new TypeMethodGroup (
+				managedName,
+				methods.OrderBy (m => m.JniName, StringComparer.Ordinal)
+					.ThenBy (m => m.JniSignature, StringComparer.Ordinal)
+					.ToList ()
+			));
+		}
+
+		foreach (var entry in dataSets.JavaToManaged) {
+			if (methodsByJavaName.ContainsKey (entry.JavaName)) {
+				continue;
+			}
+
+			methodsByJavaName [entry.JavaName] = new List<TypeMethodGroup> {
+				new TypeMethodGroup (entry.ManagedName, new List<MethodEntry> ())
+			};
+		}
+
+		return (entries, methodsByJavaName);
+	}
+
+	public static (List<TypeMapEntry> entries, Dictionary<string, List<TypeMethodGroup>> methodsByJavaName) RunNew (string[] assemblyPaths)
+	{
+		var primaryAssemblyName = Path.GetFileNameWithoutExtension (assemblyPaths [0]);
+		using var scanner = new JavaPeerScanner ();
+		var allPeers = scanner.Scan (assemblyPaths);
+		var peers = allPeers.Where (p => p.AssemblyName == primaryAssemblyName).ToList ();
+
+		var entries = peers
+			.Select (p => new TypeMapEntry (
+				p.JavaName,
+				$"{p.ManagedTypeName}, {p.AssemblyName}",
+				p.IsInterface || p.IsGenericDefinition
+			))
+			.OrderBy (e => e.JavaName, StringComparer.Ordinal)
+			.ThenBy (e => e.ManagedName, StringComparer.Ordinal)
+			.ToList ();
+
+		var methodsByJavaName = new Dictionary<string, List<TypeMethodGroup>> ();
+		foreach (var peer in peers) {
+			var managedName = $"{peer.ManagedTypeName}, {peer.AssemblyName}";
+
+			if (!methodsByJavaName.TryGetValue (peer.JavaName, out var groups)) {
+				groups = new List<TypeMethodGroup> ();
+				methodsByJavaName [peer.JavaName] = groups;
+			}
+
+			groups.Add (new TypeMethodGroup (
+				managedName,
+				peer.MarshalMethods
+					.Select (m => new MethodEntry (m.JniName, m.JniSignature, m.Connector))
+					.OrderBy (m => m.JniName, StringComparer.Ordinal)
+					.ThenBy (m => m.JniSignature, StringComparer.Ordinal)
+					.ToList ()
+			));
+		}
+
+		return (entries, methodsByJavaName);
+	}
+
+	public static string? GetCecilJavaName (TypeDefinition typeDef)
+	{
+		if (!typeDef.HasCustomAttributes) {
+			return null;
+		}
+
+		foreach (var attr in typeDef.CustomAttributes) {
+			if (attr.AttributeType.FullName != "Android.Runtime.RegisterAttribute") {
+				continue;
+			}
+
+			if (attr.ConstructorArguments.Count > 0) {
+				return ((string) attr.ConstructorArguments [0].Value).Replace ('.', '/');
+			}
+		}
+
+		return null;
+	}
+
+	public static string GetManagedName (TypeDefinition typeDef)
+	{
+		return $"{typeDef.FullName.Replace ('/', '+')}, {typeDef.Module.Assembly.Name.Name}";
+	}
+
+	static List<MethodEntry> ExtractMethodRegistrations (TypeDefinition typeDef)
+	{
+		var methods = new List<MethodEntry> ();
+
+		foreach (var method in typeDef.Methods) {
+			if (!method.HasCustomAttributes) {
+				continue;
+			}
+
+			AddRegisterMethods (method.CustomAttributes, methods);
+		}
+
+		if (typeDef.HasProperties) {
+			foreach (var prop in typeDef.Properties) {
+				if (!prop.HasCustomAttributes) {
+					continue;
+				}
+
+				AddRegisterMethods (prop.CustomAttributes, methods);
+			}
+		}
+
+		return methods;
+	}
+
+	static void AddRegisterMethods (IEnumerable<CustomAttribute> attributes, List<MethodEntry> methods)
+	{
+		foreach (var attr in attributes) {
+			if (attr.AttributeType.FullName != "Android.Runtime.RegisterAttribute" || attr.ConstructorArguments.Count < 2) {
+				continue;
+			}
+
+			var jniMethodName = (string) attr.ConstructorArguments [0].Value;
+			var jniSignature = (string) attr.ConstructorArguments [1].Value;
+			var connector = attr.ConstructorArguments.Count > 2
+				? (string) attr.ConstructorArguments [2].Value
+				: null;
+			methods.Add (new MethodEntry (jniMethodName, jniSignature, connector));
+		}
+	}
+}
